+++ init-sql/035_profile_brand_kit.sql
@@ -6,6 +6,6 @@
 ALTER TABLE `seo_site_profiles`
     ADD COLUMN `brand_palette`            JSON NULL COMMENT 'Named brand colors',
     ADD COLUMN `brand_illustration_style` TEXT NULL COMMENT 'Style hint for AI illustration prompts',
-    ADD COLUMN `brand_logo_image_id`      INT UNSIGNED NULL COMMENT 'FK seo_images.id for OG/hero logo overlay',
+    ADD COLUMN `brand_logo_image_id`      BIGINT UNSIGNED NULL COMMENT 'FK seo_images.id for OG/hero logo overlay',
     ADD CONSTRAINT `fk_profile_brand_logo`
         FOREIGN KEY (`brand_logo_image_id`) REFERENCES `seo_images`(`id`) ON DELETE SET NULL;
